Changed type from long int to int.

+++ include/arctk/debug/invariant.hpp
@@ -51,7 +51,7 @@ namespace arc //! arctk namespace
           private:
             //  -- Location --
             const std::string _file; //!< File name of the invocation location.
-            const long int    _line; //!< File line number of the invocation location.
+            const int         _line; //!< File line number of the invocation location.
             const std::string _func; //!< Function name of the invocation location.
 
             //  -- Expression --
